+++ cli/src/scheduler/task/task_create.rs
@@ -25,11 +25,11 @@ pub(crate) async fn create_task(
     mut input_json: Option<serde_json::Value>,
     remote: Vec<String>,
     metadata: Vec<String>,
-    execution_gas_price: u64,
+    execution_priority_fee_per_gas_unit: u64,
     schedule_start_ms: Option<u64>,
     schedule_start_offset_ms: Option<u64>,
     schedule_deadline_offset_ms: Option<u64>,
-    schedule_gas_price: u64,
+    schedule_priority_fee_per_gas_unit: u64,
     generator: GeneratorKind,
     gas: GasArgs,
 ) -> AnyResult<(), NexusCliError> {
@@ -109,7 +109,7 @@ pub(crate) async fn create_task(
                 None,
                 schedule_start_offset_ms,
                 schedule_deadline_offset_ms,
-                schedule_gas_price,
+                schedule_priority_fee_per_gas_unit,
                 true,
             )
             .map_err(NexusCliError::Nexus)?,
@@ -133,7 +133,7 @@ pub(crate) async fn create_task(
             entry_group: entry_group.clone(),
             input_data,
             metadata: metadata_pairs,
-            execution_gas_price,
+            execution_priority_fee_per_gas_unit,
             initial_schedule,
             generator,
         })